Move NotePreviewButton to CaseDetailsRow folder as it's only used there
