+++ src/features/CourtCaseList/CourtCaseListEntry/CaseDetailsRow/CaseDetailsRow.tsx
@@ -7,7 +7,7 @@ import ResolvedTag from "features/CourtCaseList/tags/ResolvedTag"
 import DateTime from "components/DateTime"
 import { displayedDateFormat } from "utils/formattedDate"
 import UrgentTag from "features/CourtCaseList/tags/UrgentTag"
-import { NotePreview, NotePreviewButton } from "features/CourtCaseList/NotePreviewButton"
+import { NotePreview, NotePreviewButton } from "features/CourtCaseList/CourtCaseListEntry/CaseDetailsRow/NotePreviewButton"
 import { useState } from "react"
 import groupErrorsFromReport from "utils/formatReasons/groupErrorsFromReport"
 import { filterUserNotes, getMostRecentNote } from "features/CourtCaseList/CourtCaseListEntryHelperFunction"
