+++ src/features/CourtCaseList/CourtCaseListEntry/CaseDetailsRow/NotePreviewButton.tsx
@@ -1,10 +1,10 @@
 import { truncate } from "lodash"
-import ConditionalRender from "../../components/ConditionalRender"
+import ConditionalRender from "../../../../components/ConditionalRender"
 import { Dispatch, SetStateAction } from "react"
 import { createUseStyles } from "react-jss"
-import { gdsLightGrey } from "../../utils/colours"
-import { validateMostRecentNoteDate } from "./CourtCaseListEntryHelperFunction"
-import type Note from "../../services/entities/Note"
+import { gdsLightGrey } from "../../../../utils/colours"
+import { validateMostRecentNoteDate } from "../../CourtCaseListEntryHelperFunction"
+import type Note from "../../../../services/entities/Note"
 
 const useStyles = createUseStyles({
   buttonContainer: {
